requirements.txt
pandas
plotly
streamlit
plotly.express
matplotlib.pyplot
numpy
sys